requirements.txt
asgiref==3.7.2
Django==4.2.11
pytz==2024.1
setuptools==68.1.2
six==1.16.0
sqlparse==0.4.4
supervisor==4.2.5
wheel==0.42.0
pytest==7.4.3
pytest-django==4.5.2
coverage==7.2.0
django-select2==8.4.0
django-webpush==0.3.6
